requirements.txt
aniso8601==8.0.0
blinker==1.4
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
cssselect==1.1.0
cssutils==1.0.2
Flask==1.1.1
Flask-JWT-Extended==3.24.1
flask-mongoengine==0.9.5
Flask-RESTful==0.3.7
Flask-WTF==0.14.2
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
lxml==4.4.2
MarkupSafe==1.1.1
mongoengine==0.19.1
pipenv==2018.11.26
pprint==0.1
pycparser==2.19
PyJWT==1.7.1
pymongo==3.10.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
urllib3==1.25.7
virtualenv==16.7.9
virtualenv-clone==0.5.3
Werkzeug==0.16.0
WTForms==2.2.1
psycopg2-binary==2.8.4
redis==3.4.1
